move `Option` conversions to new module & update
